fix: replace the user-liq endpoint query
